refactor(port): update UpdateBeneficiaryAccountUseCase for Long ID
Función: Puerto de entrada para la modificación de datos de una cuenta de beneficiario.
Cambios: Se actualizó el parámetro ID a tipo Long.

+++ src/main/java/org/walrex/application/port/input/UpdateBeneficiaryAccountUseCase.java
@@ -4,5 +4,5 @@ import io.smallrye.mutiny.Uni;
 import org.walrex.domain.model.BeneficiaryAccount;
 
 public interface UpdateBeneficiaryAccountUseCase {
-    Uni<BeneficiaryAccount> update(Integer id, BeneficiaryAccount beneficiaryAccount);
+    Uni<BeneficiaryAccount> update(Long id, BeneficiaryAccount beneficiaryAccount);
 }
